% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{asbcFreq}
\alias{asbcFreq}
\title{Query lemma frequencies in ASBC}
\usage{
asbcFreq(lemma, regex = FALSE)
}
\arguments{
\item{lemma}{A string. The lemma to be queried in ASBC corpus.}

\item{regex}{A string. Whether to use RegEx pattern for query.}
}
\value{
A list with each element specifying the frequency of
  occurrence of the lemma with a particular Part-Of-Speech tag.
}
\description{
Query lemma frequencies in ASBC
}
