% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{install_db}
\alias{install_db}
\title{Install databases from local file paths or urls}
\usage{
install_db(
  cwn_path = NULL,
  db_path = NULL,
  asbc_freq_path = NULL,
  install = c(cwn = TRUE, db = TRUE, asbc_freq = TRUE)
)
}
\arguments{
\item{cwn_path}{A string. The path to \code{CwnGraph.zip}.
Defaults to \code{NULL}, which downloads \code{CwnGraph.zip}
from the internet and installs it to \code{lexicoR}.}

\item{db_path}{A string. The path to \code{database.rda} or
\code{database.zip}. Defaults to \code{NULL}, which
downloads \code{database.zip} from the internet and
installs it to \code{lexicoR}.}

\item{install}{Logical. A logical vector specifying which
databases to install.}
}
\value{
The path to the directory where the databases
  were installed.
}
\description{
Install databases from local file paths or urls
}
