% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{pttFreq}
\alias{pttFreq}
\title{Query lemma frequencies on PTT}
\usage{
pttFreq(lemma, board, year = 2019)
}
\arguments{
\item{lemma}{A string. The lemma to be queried on PTT.}

\item{board}{A string. The board on PTT to search for.}

\item{year}{Integer. The year to search for.}
}
\value{
A list with each element specifying the frequency of
  occurrence of the lemma with a particular Part-Of-Speech tag.
}
\description{
Query lemma frequencies on PTT
}
