% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query.lu}
\alias{query.lu}
\title{Query lexical databases}
\usage{
query.lu(
  lu,
  isSimp = FALSE,
  regex = FALSE,
  db = c("deeplex", "cld"),
  as_tibble = TRUE
)
}
\arguments{
\item{lu}{A string. The lexical unit to be queried in the databases.}

\item{isSimp}{Logical. Whether \code{lu} is in simplified Chinese.
Defaults to \code{FALSE}, which treat \code{lu} as traditional Chinese.}

\item{regex}{Logical. Whether to use a regular expression pattern in
the argument \code{lu} for searching. Defaults to \code{FALSE},
which searches for exact matches.}

\item{db}{Character. The databases to be searched. Currently only
\code{deeplex} and \code{cld} (Chinese Lexical Database) are available.}

\item{as_tibble}{Logical. Whether to convert the query results to
a \code{tibble} (a special type of \code{data.frame} provided by
the package \pkg{tibble}). Defaults to \code{TRUE}, which makes the data
frame printed out in the console to have a nicer look.}
}
\value{
A data frame (\code{tibble}).
}
\description{
Query lexical databases
}
