% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_local}
\alias{query_local}
\title{Query lexical data stored in local databases.}
\usage{
query_local(query, regex = FALSE, db = c("deeplex", "cld"), as_tibble = TRUE)
}
\arguments{
\item{query}{A string. A string that matches the words (or lexical units)
in the databases.}

\item{regex}{Logical. Whether to use a regular expression pattern in
the argument \code{query}. Defaults to \code{FALSE}, which searches
for exact matches.}

\item{db}{Character. The databases to be searched. Currently only
\code{deeplex} and \code{cld} (Chinese Lexical Database) are available.}

\item{as_tibble}{Logical. Whether to convert the query results to
a \code{tibble} (a special type of \code{data.frame} provided by
the package \pkg{tibble}). Defaults to \code{TRUE}, which makes the data
frame printed out in the console to have a nicer look.}
}
\value{
A data frame (\code{tibble}).
}
\description{
Query lexical data stored in local databases.
}
